Updated to v067r23 release.
byuu says:

Fixed bsnes launcher on Windows XP
Fixed Windows bsnes launcher internationalization support (emulator can
be in a folder with spaces and Japanese characters, and you can drag
a Japanese file name onto the launcher, and it will load it properly)
Moved fast CPU to use a switch table for MMIO, unfortunately for no
speed gain
Bus::read/write take uint24 parameters for address, luckily no speed
penalty
MMIOAccess gained a handle() function, and hid the mmio[] table. Makes
hooking it cleaner
Added malloc.h header to nall/function.hpp to fix a ridiculous GCC 4.5.0
error
Fixed a fairly large bug in the fast CPU IRQ handler, which fixes
Robocop et al
Forgot to bump revision to .24 in the compiled binaries, too lazy to
recompile or hex edit to change them

Unfortunately, in order to add nice battery usage, I have to add the
sleep calls to the video and audio wait loops. But they don't know
anything about the GUI and its settings, nor do I really want to make
them know about this setting. I do not want to force allow it. Even with
the media timer trick, Sleep(0) makes Vsync+Async fail a lot more
frequently than never sleeping at all. I would rather laptop users
suffer 100% utilization of a single core than for all users to not be
able to get good audio+video sync. Not sure what to do about that, so
I'll probably just remove the battery usage comment from performance
mode for now.

+++ snes/memory/memory-inline.hpp
@@ -51,7 +51,7 @@ MappedRAM::MappedRAM() : data_(0), size_(-1U), write_protect_(false) {}
 
 //Bus
 
-uint8 Bus::read(unsigned addr) {
+uint8 Bus::read(uint24 addr) {
   #if defined(CHEAT_SYSTEM)
   if(cheat.active() && cheat.exists(addr)) {
     uint8 r;
@@ -63,7 +63,7 @@ uint8 Bus::read(unsigned addr) {
   return p.access->read(p.offset + addr);
 }
 
-void Bus::write(unsigned addr, uint8 data) {
+void Bus::write(uint24 addr, uint8 data) {
   Page &p = page[addr >> 8];
   return p.access->write(p.offset + addr, data);
 }
